# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для студенческой платформы Центра поддержки и развития молодёжи",
        "title": "Student Platform API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@studentplatform.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/grades/my": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список всех оценок текущего пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grades"
                ],
                "summary": "Получение всех оценок текущего студента",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/grade.GradeResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/grades/my/courses/{courseId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает оценки текущего студента по конкретному курсу",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grades"
                ],
                "summary": "Получение оценок по курсу",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course ID",
                        "name": "courseId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/grade.GradeResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/grades/my/period": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает оценки текущего студента за указанный период",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grades"
                ],
                "summary": "Получение оценок за период",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Начало периода (RFC3339)",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Конец периода (RFC3339)",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/grade.GradeResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/grades/my/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает общую сводку по успеваемости студента",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grades"
                ],
                "summary": "Получение сводки успеваемости",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/grade.StudentSummary"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/grades/students/{studentId}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создаёт новую оценку для студента",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grades"
                ],
                "summary": "Создание новой оценки (для преподавателей)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student ID",
                        "name": "studentId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные оценки",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/grade.CreateGradeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/grade.GradeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/grades/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаляет оценку",
                "tags": [
                    "grades"
                ],
                "summary": "Удаление оценки",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Grade ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет существующую оценку",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "grades"
                ],
                "summary": "Обновление оценки",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Grade ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/grade.UpdateGradeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/grade.GradeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "grade.CourseSummary": {
            "type": "object",
            "properties": {
                "average": {
                    "description": "средняя оценка",
                    "type": "number"
                },
                "course_id": {
                    "type": "string"
                },
                "course_name": {
                    "type": "string"
                },
                "grades_count": {
                    "description": "количество оценок",
                    "type": "integer"
                },
                "total_credits": {
                    "description": "всего кредитов",
                    "type": "integer"
                }
            }
        },
        "grade.CreateGradeRequest": {
            "type": "object",
            "required": [
                "course_id",
                "date",
                "max_value",
                "type",
                "value"
            ],
            "properties": {
                "comment": {
                    "type": "string"
                },
                "course_id": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "max_value": {
                    "type": "number",
                    "minimum": 1
                },
                "type": {
                    "enum": [
                        "exam",
                        "test",
                        "homework",
                        "project",
                        "activity"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/grade.GradeType"
                        }
                    ]
                },
                "value": {
                    "type": "number",
                    "minimum": 0
                },
                "weight": {
                    "type": "number",
                    "minimum": 0
                }
            }
        },
        "grade.GradeResponse": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "course_id": {
                    "type": "string"
                },
                "course_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "max_value": {
                    "type": "number"
                },
                "type": {
                    "$ref": "#/definitions/grade.GradeType"
                },
                "value": {
                    "type": "number"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "grade.GradeType": {
            "type": "string",
            "enum": [
                "exam",
                "test",
                "homework",
                "project",
                "activity"
            ],
            "x-enum-comments": {
                "GradeTypeActivity": "активность на занятии",
                "GradeTypeExam": "экзамен",
                "GradeTypeHomework": "домашнее задание",
                "GradeTypeProject": "проект",
                "GradeTypeTest": "зачёт"
            },
            "x-enum-descriptions": [
                "экзамен",
                "зачёт",
                "домашнее задание",
                "проект",
                "активность на занятии"
            ],
            "x-enum-varnames": [
                "GradeTypeExam",
                "GradeTypeTest",
                "GradeTypeHomework",
                "GradeTypeProject",
                "GradeTypeActivity"
            ]
        },
        "grade.StudentSummary": {
            "type": "object",
            "properties": {
                "courses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/grade.CourseSummary"
                    }
                },
                "last_updated": {
                    "type": "string"
                },
                "overall_average": {
                    "description": "общий средний балл",
                    "type": "number"
                },
                "total_credits": {
                    "description": "всего кредитов",
                    "type": "integer"
                }
            }
        },
        "grade.UpdateGradeRequest": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}